requirements.txt
Django==2.0
django-recaptcha==0.0.4
django-simple-captcha==0.5.10
django-widget-tweaks==1.4.3
django-localflavor==2.0
django-starfield
Pillow
python-decouple==3.1
six==1.12.0
django-multiselectfield==0.1.8
django-registration==3.0
pytz==2018.9
django-ranged-response==0.2.0
confusable-homoglyphs==3.2.0
django-extensions==2.1.6
requests==2.21.0